requirements.txt
customtkinter>=5.2.2
aiohttp>=3.9.3
Pillow>=10.2.0
